pyproject.toml
[tool.poetry]
name = "v3-competitor-analysis"
version = "0.1.0"
description = ""
authors = ["tuannd4 <tuannd4@smartosc.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.1.0"
openai-agents = "^0.0.7"
asyncio = "^3.4.3"
rich = "^14.0.0"
markdown = "^3.7"
weasyprint = "^65.0"
streamlit = "^1.44.1"
streamlit-tags = "^1.2.8"
matplotlib = "^3.10.3"
seaborn = "^0.13.2"
requests = "^2.32.3"
beautifulsoup4 = "^4.13.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
